Second attempt at fixing warning on Windows.
R=asiva@google.com

Review URL: https://codereview.chromium.org//616103004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@40827 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/object.cc
@@ -16735,7 +16735,7 @@ uint32_t Bigint::AsTruncatedUint32Value() const {
   const intptr_t used = Used();
   if (used == 0) return 0;
   const uint32_t digit0 = DigitAt(0);
-  return Neg() ? static_cast<uint32_t>(-digit0) : digit0;
+  return Neg() ? static_cast<uint32_t>(-static_cast<int32_t>(digit0)) : digit0;
 }
